:sparkles: Added a task that says something
More less simple tasks that can say predefined messages to a channel
This is to mimic some functionality I had in ruby using webhooks that I
wanted to port over.

Not all the messages have been ported and I feel the code is not super
amazing, but Im super tired at this point and it was good practice
nontheless.

Also this is the last commit Ill do without adding tests. I think I got
the hang of it at this point.